♻️ ApiValidatedField example, default imports
