+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { lazy, Suspense, useMemo } from 'react'
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 //import { Test as ComponentTest } from './components/TextLoader'
-import { Example as ComponentTest } from './components/IBANField.example'
+import { Example as ComponentTest } from './components/ApiValidatedField.example'
 
 import {
   createTheme,
